Refactor everything
When I tried using the package, it didn't work out of the
box (probably bit-rot). It's ideas, however, were solid.  So, I
cleaned things up and in the process basically refactored everything.

Some great efforts are taken to format the results so that they match
typical Org results. All the source evaluation functions Org
provides (that I could find and understand) return a plain string. But
results are not presented to users as plain strings. Instead they are
prefixed with colons, formatted as a table, et cetera. To cope with
this (and my limited time and patience), I pass the plain text results
through a dummy source block in hopes that it will format the results
correctly (it doesn't) and copy that output.  The output formatting is
correct for the ":results output" header, but not for all headers,
like "table". This might be fixed by figuring out how to pass
parameters and info arguments to elisp calls to the org execute
related functions.  This commit, like previous versions, cannot work
with sessions.